Some minor documentaion and notes added
I guess the project is complete so far. I`ve implemented everything I needed. Adding more functionality will create an unnecessary bloat and `blur` the purpose of this utility. It was a very exciting and useful project for me. I`m not abandoning it, but I don`t think that I`ll need some new feature in the near future